Renamed the utility package to util for consistency with many other projects.
