+++ src/test/java/org/TeamCodeDefy/util/BookConversionTest.java
@@ -3,7 +3,6 @@ package org.TeamCodeDefy.util;
 import org.TeamCodeDefy.entities.Book;
 import org.TeamCodeDefy.googleBooksApi.GoogleBookResponse;
 import org.TeamCodeDefy.persistance.GoogleBooksApiDao;
-import org.TeamCodeDefy.utilities.BookConversion;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
